requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Airflow and its Snowflake provider (ensure Airflow is >=2.9.0)
apache-airflow
apache-airflow-providers-snowflake
astronomer-cosmos
snowflake-connector-python
snowflake-sqlalchemy
pandas
numpy
pyarrow
nba_api
scikit-learn
matplotlib